debug share post et delete post STEP

+++ server/controllers/post/deletePost.js
@@ -4,37 +4,6 @@ const fs = require("fs");
 const ApiError = require("../../error/ApiError");
 const { Socket } = require("../../utils/socket");
 
-function deleteSharedPostsData(req, res, next, post) {
-    //Si le post a été partagé le tableau usersShared contient les id des utilisateurs ayant partagé
-    if (post.usersShared.length > 0) {
-        post.usersShared.map((ID) =>
-            PostModel.find({
-                userId: ID,
-                sharedPostId: req.params.id,
-            })
-                .then((sharedPosts) => {
-                    sharedPosts.map((sharedPost) => {
-                        sharedPost.sharedImage = "";
-                        sharedPost.sharedTexte =
-                            "La publication a été supprimée";
-                        PostModel.updateOne({ _id: sharedPost._id }, sharedPost)
-                            .then(() => deleteSharedPostsData(sharedPost))
-                            .then(() =>
-                                Socket.emit("likeAndLovesResponse", sharedPost)
-                            )
-                            .catch((error) => {
-                                return next(ApiError.badRequest(error.message));
-                            });
-                    });
-                })
-                .catch((error) => {
-                    return next(ApiError.notFound(error.message));
-                })
-        );
-    }
-    return post;
-}
-
 /**
  * * deletePost :
  * Fonction pour supprimer (delete) un post de la BDD,
@@ -54,8 +23,8 @@ const deletePost = (req, res, next) => {
             post.userId != req.session.userId &&
                 next(ApiError.unauthorized("Accès refusé"));
             //Si c'est un post partagé alors sharedPostId contient l'id du post d'origine
-            if (post.sharedPostId) {
-                PostModel.findOne({ _id: req.params.id })
+            if (post.sharedPostId && post.fromPostId === post.sharedPostId) {
+                PostModel.findOne({ _id: post.sharedPostId })
                     .then((originalPost) => {
                         const userIndex = originalPost.usersShared.indexOf(
                             req.session.userId
@@ -68,19 +37,78 @@ const deletePost = (req, res, next) => {
                         return originalPost;
                     })
                     .then((originalPost) => {
+                        const obj = {
+                            originalPostId: originalPost._id,
+                            userId: req.session.userId,
+                        };
                         PostModel.updateOne(
                             { _id: originalPost._id },
                             originalPost
                         )
-                            .then(() =>
-                                Socket.emit(
-                                    "likeAndLovesResponse",
-                                    originalPost
-                                )
-                            )
+                            .then(() => Socket.emit("shareDeleted", obj))
                             .catch((error) => {
                                 return next(ApiError.badRequest(error.message));
                             });
+                    })
+                    .catch(() => {
+                        return post;
+                    });
+            } else if (post.fromPostId) {
+                PostModel.findOne({ _id: post.fromPostId })
+                    .then((fromPost) => {
+                        const userIndex = fromPost.usersShared.indexOf(
+                            req.session.userId
+                        );
+                        userIndex !== -1 &&
+                            fromPost.usersShared.splice(userIndex, 1);
+                        userIndex !== -1 &&
+                            fromPost.shares > 0 &&
+                            (fromPost.shares -= 1);
+                        return fromPost;
+                    })
+                    .then((fromPost) => {
+                        const obj = {
+                            originalPostId: fromPost._id,
+                            userId: req.session.userId,
+                        };
+                        PostModel.updateOne({ _id: fromPost._id }, fromPost)
+                            .then(() => Socket.emit("shareDeleted", obj))
+                            .catch((error) => {
+                                return next(ApiError.badRequest(error.message));
+                            });
+                    })
+                    .catch(() => {
+                        return post;
+                    });
+            }
+            return post;
+        })
+        .then((post) => {
+            if (post && post.usersShared.length > 0) {
+                PostModel.find({
+                    sharedPostId: req.params.id,
+                })
+                    .then((sharedPosts) => {
+                        sharedPosts.map((sharedPost) => {
+                            sharedPost.sharedImage = "";
+                            sharedPost.sharedTexte =
+                                "La publication a été supprimée";
+                            PostModel.updateOne(
+                                { _id: sharedPost._id },
+                                sharedPost
+                            )
+                                .then(() =>
+                                    Socket.emit(
+                                        "PropageContentDelete",
+                                        post._id
+                                    )
+                                )
+                                .catch((error) => {
+                                    return next(
+                                        ApiError.badRequest(error.message)
+                                    );
+                                });
+                        });
                     })
                     .catch((error) => {
                         return next(ApiError.notFound(error.message));
@@ -89,44 +117,8 @@ const deletePost = (req, res, next) => {
             return post;
         })
         .then((post) => {
-            if (post.usersShared.length > 0) {
-                post.usersShared.map((ID) =>
-                    PostModel.find({
-                        userId: ID,
-                        sharedPostId: req.params.id,
-                    })
-                        .then((sharedPosts) => {
-                            sharedPosts.map((sharedPost) => {
-                                sharedPost.sharedImage = "";
-                                sharedPost.sharedTexte =
-                                    "La publication a été supprimée";
-                                PostModel.updateOne(
-                                    { _id: sharedPost._id },
-                                    sharedPost
-                                )
-                                    .then(() =>
-                                        Socket.emit(
-                                            "likeAndLovesResponse",
-                                            sharedPost
-                                        )
-                                    )
-                                    .catch((error) => {
-                                        return next(
-                                            ApiError.badRequest(error.message)
-                                        );
-                                    });
-                            });
-                        })
-                        .catch((error) => {
-                            return next(ApiError.notFound(error.message));
-                        })
-                );
-            }
-            return post;
-        })
-        .then((post) => {
             //S'il y a des commentaires
-            if (post.commentaires.length > 0) {
+            if (post && post.commentaires.length > 0) {
                 CommentModel.deleteMany({ postId: post._id }).catch((error) => {
                     return next(ApiError.notFound(error.message));
                 });
@@ -134,13 +126,28 @@ const deletePost = (req, res, next) => {
             return post;
         })
         .then((post) => {
-            setTimeout(() => {
-                if (post.image) {
-                    const filename = post.image.split("/images/posts/")[1];
-                    fs.unlink(`images/posts/${filename}`, () => {
-                        PostModel.deleteOne({ _id: req.params.id })
+            post &&
+                setTimeout(() => {
+                    if (post.image) {
+                        const filename = post.image.split("/images/posts/")[1];
+                        fs.unlink(`images/posts/${filename}`, () => {
+                            PostModel.deleteOne({ _id: post._id })
+                                .then(() => {
+                                    Socket.emit("postDeleted", post._id);
+                                    return res.status(200).json({
+                                        message: "Post supprimée !",
+                                    });
+                                })
+                                .catch((error) => {
+                                    return next(
+                                        ApiError.internal(error.message)
+                                    );
+                                });
+                        });
+                    } else {
+                        PostModel.deleteOne({ _id: post._id })
                             .then(() => {
-                                Socket.emit("postDeleted", req.params.id);
+                                Socket.emit("postDeleted", post._id);
                                 return res.status(200).json({
                                     message: "Post supprimée !",
                                 });
@@ -148,20 +155,8 @@ const deletePost = (req, res, next) => {
                             .catch((error) => {
                                 return next(ApiError.internal(error.message));
                             });
-                    });
-                } else {
-                    PostModel.deleteOne({ _id: req.params.id })
-                        .then(() => {
-                            Socket.emit("postDeleted", req.params.id);
-                            return res.status(200).json({
-                                message: "Post supprimée !",
-                            });
-                        })
-                        .catch((error) => {
-                            return next(ApiError.internal(error.message));
-                        });
-                }
-            }, 500);
+                    }
+                }, 500);
         })
         .catch((error) => {
             return next(ApiError.notFound(error.message));
